requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dateutil==2.8.2
asyncio==3.4.3